split Models.kt, remove getIdentityRecord method
