+++ cordlib-contracts-states/src/main/kotlin/com/luxoft/blockchainlab/corda/hyperledger/indy/data/state/IndyCredential.kt
@@ -1,8 +1,8 @@
 package com.luxoft.blockchainlab.corda.hyperledger.indy.data.state
 
 import com.luxoft.blockchainlab.corda.hyperledger.indy.data.schema.CredentialSchemaV1
-import com.luxoft.blockchainlab.hyperledger.indy.CredentialInfo
-import com.luxoft.blockchainlab.hyperledger.indy.CredentialRequestInfo
+import com.luxoft.blockchainlab.hyperledger.indy.models.CredentialInfo
+import com.luxoft.blockchainlab.hyperledger.indy.models.CredentialRequestInfo
 import net.corda.core.contracts.LinearState
 import net.corda.core.contracts.UniqueIdentifier
 import net.corda.core.identity.AbstractParty
