+++ cordlib/src/main/kotlin/com.luxoft.blockchainlab.corda.hyperledger.indy/flow/b2b/GetDidFlowB2B.kt
@@ -3,8 +3,7 @@ package com.luxoft.blockchainlab.corda.hyperledger.indy.flow.b2b
 import co.paralleluniverse.fibers.Suspendable
 import com.luxoft.blockchainlab.corda.hyperledger.indy.flow.indyUser
 import com.luxoft.blockchainlab.corda.hyperledger.indy.flow.whoIs
-import com.luxoft.blockchainlab.hyperledger.indy.IdentityDetails
-import com.luxoft.blockchainlab.hyperledger.indy.roles.getIdentity
+import com.luxoft.blockchainlab.hyperledger.indy.models.IdentityDetails
 import com.luxoft.blockchainlab.hyperledger.indy.utils.SerializationUtils
 import net.corda.core.flows.*
 import net.corda.core.identity.CordaX500Name
@@ -51,7 +50,9 @@ object GetDidFlowB2B {
         @Suspendable
         override fun call() {
             try {
-                flowSession.send(indyUser().getIdentity().getIdentityRecord())
+                val serializedIdentity = SerializationUtils.anyToJSON(indyUser().getIdentity())
+
+                flowSession.send(serializedIdentity)
             } catch (e: Exception) {
                 logger.error("", e)
                 throw FlowException(e.message)
