+++ cordlib/src/main/kotlin/com.luxoft.blockchainlab.corda.hyperledger.indy/flow/b2c/VerifyCredentialFlowB2C.kt
@@ -8,10 +8,10 @@ import com.luxoft.blockchainlab.corda.hyperledger.indy.flow.ProofPredicate
 import com.luxoft.blockchainlab.corda.hyperledger.indy.flow.indyUser
 import com.luxoft.blockchainlab.corda.hyperledger.indy.flow.whoIsNotary
 import com.luxoft.blockchainlab.corda.hyperledger.indy.service.connectionService
-import com.luxoft.blockchainlab.hyperledger.indy.CredentialFieldReference
-import com.luxoft.blockchainlab.hyperledger.indy.CredentialPredicate
+import com.luxoft.blockchainlab.hyperledger.indy.models.CredentialFieldReference
+import com.luxoft.blockchainlab.hyperledger.indy.models.CredentialPredicate
 import com.luxoft.blockchainlab.hyperledger.indy.IndyUser
-import com.luxoft.blockchainlab.hyperledger.indy.Interval
+import com.luxoft.blockchainlab.hyperledger.indy.models.Interval
 import net.corda.core.contracts.Command
 import net.corda.core.contracts.StateAndContract
 import net.corda.core.flows.*
