+++ indy-utils/src/main/kotlin/com/luxoft/blockchainlab/hyperledger/indy/LedgerService.kt
@@ -1,5 +1,6 @@
 package com.luxoft.blockchainlab.hyperledger.indy
 
+import com.luxoft.blockchainlab.hyperledger.indy.models.*
 import com.luxoft.blockchainlab.hyperledger.indy.utils.SerializationUtils
 import mu.KotlinLogging
 import org.hyperledger.indy.sdk.ledger.Ledger
