+++ indy-utils/src/main/kotlin/com/luxoft/blockchainlab/hyperledger/indy/models/Credential.kt
@@ -1,0 +1,154 @@
+package com.luxoft.blockchainlab.hyperledger.indy.models
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+/**
+ * Represents credential offer structure from.
+ *
+ * Example:
+ * {
+ *  "schema_id":"V4SGRU86Z58d6TV7PBUe6f:2:gvt:1.0",
+ *  "cred_def_id":"V4SGRU86Z58d6TV7PBUe6f:3:CL:11:TAG_1",
+ *  "key_correctness_proof":{
+ *      "c":"13761754840304977622723160610073451825885870027886575679644893270880210242048",
+ *      "xz_cap":"340594989055020773127826593833336513921518724477036785882482817268296213948251287179407974588164919084334695651740569687300632847210154572884165159121599184025645166277098212790056340367675548818550765726296311292965992912074551019256115749936791678223911283791007656125545683865408985565241327271094579434170968145628655514446630544302951035920170794170246151362177816426545156184025449231122642094985676281843584455651701883820074327331609796847757062117491672731024235304103252859539661490806446048072318460507557965757317583074214165065297303144649946568723405884817293591612762145863174650181663890220148408024636352402091756427859989263831219418429810184080686385298141705690561349412826",
+ *      "xr_cap":[
+ *          [
+ *              "sex",
+ *              "114901712389925570747308755505959234181050635018455439015518673538665420714600742391112487936318767337811364541526998298909351153769381909405458975688912346829659002463105330151270376906904767852191120948221184076757548668178839902476764369158701651586213020361657048021989142635432083872470012807575132413533225131436609909265898568850849369239607732687307503724259039320052575044915702202127827497635261940052132289706849314498118894709987533054997486796495324835894021134546049747254662911845883745637179643031937372509664499727668214235022494858047366785583118269238559014142100466017577191051916007450284436475900646482814224888782960060518024693146673531063922330415915936725507733741766"
+ *          ],
+ *          [
+ *              "height",
+ *              "87925827250479498410024054128993506647408563496235591997576250953229036920959413598053360534336792807215061873972469226074015743490418917494435585457616775963853677170038454238754262365048434708885925169321027257392523529242823878080721156710414911339886319487461055502252343779360662670429306854831019261366623391128187478643887112210033870842054849023198347212729401623194059443628755320835685092944707226129817372493915681098131332323544477428002628840909180360070413137657346950550327191998701401766442452826340386823830760371963355720574367910987941501327853296595241416357764706345537910660613428363744193215240447807617629842317838672093122113424390775085560952953653914862062733460343"
+ *          ],
+ *          [
+ *              "age",
+ *              "293580247693151878814398198943884612452144653998669902194814683694462974264538931634280983530503958470285180952289780325087586570844097897587837512835758168930990380054580984461202240505574024403678079197953977922329186677771110248003920244900099391489674646294554422537185554996194781678780675220797160151412165442391881155725667489570391825773712503404857486672042567138673752180271219842671234810506725182196822255079859018162540163053327438979514404327868768200699692081177648933881745792483704320534292224956976674666455897176640767709821806484479564706964601579894630931579340752620471333071445922804710427962520331104667596003689846861532048740977647311084634223124206852506298920649196"
+ *          ],
+ *          [
+ *              "name",
+ *              "151838209223886649926918852654347303937974570263808192853823982025169602769061906064239685663594882187900023324484167008090681531752389241462722998187033577222372389532314754276217632204577207156930548234292944901257640439529710798080879389939943720570224719913036311420119996340541020890815695086382900217327755857286037846480798201129396307170715168897386494499064864127724739789806487609779050080617809107250172341879674283079820670267527232446397367776741773774314369698353651749919265249180651439166600686556843514441308902088097745692987067106281950141851140402550212265361047132303049470330028968862688869371904469006037728977048516991848156345476652473561273562256067871557885766170714"
+ *          ]
+ *      ]
+ *  },
+ *  "nonce":"107428647282355717425385"
+ * }
+ */
+data class CredentialOffer(
+    override val schemaId: String,
+    @JsonProperty("cred_def_id") override val credentialDefinitionId: String,
+    val keyCorrectnessProof: KeyCorrectnessProof,
+    val nonce: String
+) : ContainsSchemaId, ContainsCredentialDefinitionId
+
+data class KeyCorrectnessProof(val c: String, val xzCap: String, val xrCap: List<List<String>>)
+
+/**
+ * Represents credential
+ *
+ * Example:
+ * {
+ *  "schema_id":"V4SGRU86Z58d6TV7PBUe6f:2:gvt:1.0",
+ *  "cred_def_id":"V4SGRU86Z58d6TV7PBUe6f:3:CL:11:TAG_1",
+ *  "rev_reg_id":null,
+ *  "values":{
+ *      "height":{
+ *          "raw":"175",
+ *          "encoded":"175"
+ *      },
+ *      "age":{
+ *          "raw":"28",
+ *          "encoded":"28"
+ *      },
+ *      "sex":{
+ *          "raw":"male",
+ *          "encoded":"5944657099558967239210949258394887428692050081607692519917050"
+ *      },
+ *      "name":{
+ *          "raw":"Alex",
+ *          "encoded":"1139481716457488690172217916278103335"
+ *      }
+ *   },
+ *   "signature":{
+ *      "p_credential":{
+ *          "m_2":"49667387409072791895333303899794250195621079019939037173941098352432267351052",
+ *          "a":"46472196767765770907394666564914908827636941437774002244808957817929335484901198469847816191888793787424454173615251713617437642140800957095444096953289022795933584947973882912549573017031326042172023742700124126097617123527165994497561684340831487031052756596925838552154440928523436704804747409373866015708822951087319456147879379870504295076305153309850751921996261053799107481550379520843249944485325939052177624821394027500110005624266210467198373597250503566237828450667266448183418227918400401704699885565245770811435243112403582975529471329394083897765235108123969958207741673911072934717461483113545266188644",
+ *          "e":"259344723055062059907025491480697571938277889515152306249728583105665800713306759149981690559193987143012367913206299323899696942213235956742929778868433842729916830526494323897253",
+ *          "v":"9338887988426201988765421562196770167154879800148014814333922180246755157926098177887019760590189644706588962889376430933272905256523280844850525307522972101041395604753489298645400791561173204918580803173446183370402135655019581889686304488882920336043712491329845108393523901867723835284305334314940293181692913962879688500165934035632199449508262321891885131213286008555035612673560763018079383317413701517586293134534116818845722224841680073553237282409082729185652870454427020765139523137039257689447228666527191123526642595385818366411298045778216509692039175062669997796006718498422877412133922090518369374952914510984290167002861717430237148314299826063537846309857330538893275349996073101096144565363217117483025547405594356909440065486323990781381766415423430083890878386623400226799579298704315850982500754033"
+ *      },
+ *      "r_credential":null
+ *   },
+ *   "signature_correctness_proof":{
+ *      "se":"1284226342082548650319469664134606254016018076595188283130710346821644123498040330813630022863109638947088616640037389617195067147163300020258649281610103910536383626344523082298355692748130628416477490704057208576461724003588709297793772889042396317116869584149262342158061174817847081593704495774591524436555209826229614130157263249266720446677213266380692272756829476405786651218028670075213938724635127001673178988057648805807845178755090432507569406752220158843852268144659430034286675620223890177350461149637114343023508452878051585971310572205236878421406629892019776501637867344989734772211374366088132166423",
+ *      "c":"27921508545451359693593235318800947503350694274033708495199465258660793720689"
+ *   },
+ *   "rev_reg":null,
+ *   "witness":null
+ *  }
+ */
+data class Credential(
+    override val schemaId: String,
+    @JsonProperty("cred_def_id") override val credentialDefinitionId: String,
+    @JsonProperty("rev_reg") val revocationRegistry: RawJsonMap?,
+    val witness: RawJsonMap?,
+    @JsonProperty("rev_reg_id") override val revocationRegistryId: String?,
+    val values: Map<String, CredentialValue>,
+    val signature: Map<String, RawJsonMap?>,
+    val signatureCorrectnessProof: RawJsonMap
+) : ContainsSchemaId, ContainsCredentialDefinitionId, ContainsRevocationRegistryId
+
+data class CredentialValue(val raw: String, val encoded: String)
+
+data class CredentialInfo(val credential: Credential, val credRevocId: String?, val revocRegDeltaJson: String?)
+
+/**
+ * Represents credential request
+ *
+ * Example:
+ * {
+ *  "prover_did":"V4SGRU86Z58d6TV7PBUe6f",
+ *  "cred_def_id":"V4SGRU86Z58d6TV7PBUe6f:3:CL:11:TAG_1",
+ *  "blinded_ms":{
+ *      "u":"72450058149003836603428182664797985470713819794261103126480867221755083585000019083805880372220123137259404481119363367548533297964699411822665134867264734782748083131245644644829379611885825013696986984794316038880504276824857600130124722761268165098283024899944184232804885740644223950687056600506599750855615763905035549112546265746441459321427491710929699552131014192618377018001144050499899842399856735232631240922205358342612967266348360240558563773073817892531681534248845102456387131736087786422967038955678533370053125605715326022828981184036662477668484559656286399716565826498183240073734557435082596734065",
+ *      "ur":null
+ *   },
+ *   "blinded_ms_correctness_proof":{
+ *      "c":"31090608731660042115240737510118668008332292387792553721852538017836016343046",
+ *      "v_dash_cap":"621718668163798225811486019897029182898804769156507569494749044241955862304516215318245510489009046007188401984864804588528641054499552984506583720030179227648767916980985982183552124476085689924812679856949430827708131983941887148948972968983789570929564392534906650079057667375456353133856194521441614464136439297585036819776218427302870626531718509167391460209553569722115239275775660807948979706555762042625362148149841737624635909956272979193386969558740798629192930114777368203501554600798530123840820679251140249079971589888320217707423350987128237788679490998464629247139577687505921078733467007002695480111134585335699005335287893428608708147970609552011478080930963461588079631276592063200753226925160329311",
+ *      "ms_cap":"32160747770865816463105278561268942103905179895830400820850731916868271052319805671289486219845889897290386865499398383995979398737576225013066455933796815223495239463836014358141"
+ *   },
+ *   "nonce":"1186917666793748761028490"
+ * }
+ *
+ */
+data class CredentialRequestInfo(
+    val request: CredentialRequest,
+    val metadata: CredentialRequestMetadata
+)
+
+data class CredentialRequest(
+    val proverDid: String,
+    @JsonProperty("cred_def_id") override val credentialDefinitionId: String,
+    val blindedMs: RawJsonMap,
+    val blindedMsCorrectnessProof: RawJsonMap,
+    val nonce: String
+) : ContainsCredentialDefinitionId
+
+/**
+ * Represents credential request metadata
+ *
+ * Example:
+ * {
+ *  "master_secret_blinding_data":{
+ *      "v_prime":"19996992452923335739444111602323353146400610615332675413132845150786968475048282265691890632536001360833725249574222779467320862724682652038419474915879533807321016579521760714569205647449244233907560466688603170418829351681403404935647974607945749195678658844745904385566324446192755479942309264166454465738364166358492528736414247799673112394278935683056028326248487491960209103589894282360872956217377481736453946808616539396839604132182798890680363804230266296163940604353270837853010999758985713684313386130790502196378124744035357918753867699384331285330409134810007357798671352898872667892627983510765893834524972207798318482093594850",
+ *      "vr_prime":null
+ *  },
+ *  "nonce":"1186917666793748761028490",
+ *  "master_secret_name":"masterSecretId"
+ * }
+ */
+data class CredentialRequestMetadata(
+    val masterSecretBlindingData: RawJsonMap,
+    val masterSecretName: String,
+    val nonce: String
+)
