+++ indy-utils/src/main/kotlin/com/luxoft/blockchainlab/hyperledger/indy/models/CredentialDefinition.kt
@@ -1,0 +1,97 @@
+package com.luxoft.blockchainlab.hyperledger.indy.models
+
+import com.fasterxml.jackson.annotation.JsonIgnore
+import com.fasterxml.jackson.annotation.JsonProperty
+
+/**
+ * Represents credential definition
+ *
+ * Example:
+ * {
+ *  "ver":"1.0",
+ *  "id":"V4SGRU86Z58d6TV7PBUe6f:3:CL:11:TAG_1",
+ *  "schemaId":"11",
+ *  "type":"CL",
+ *  "tag":"TAG_1",
+ *  "value":{
+ *      "primary":{
+ *          "n":"97825789590062692150319390633467047046201095546380339364160961030429137551345980953797290715661332171822654941232742526180523305975436127060967945697600291470349688931509035153112372164175391196321435197412518123214148508985637308223324599142177695725198704394816353613227336593431540982906270549568391784047827941399905911323367262453292590513881444395762311018337780680352564520585183115243471839552585387700359266104870621499920269868340747617483585174509821630813967372030256014647503731893474943716949401787092423796309347202300689031224683810152043236004828131092812383226231107878482692717606522179435737963809",
+ *          "s":"39733411017179560244024968899642664398180361222127496719414627387111951836383304903709718151952941019469808679723550262854530221428192476695197984887679496172637317619920757202465712652277595749798862290700699750698253379509815941232001323375550623145841707962772398431536953095683812602550974096101709249765431308450233784849698817513922831987542597248666167913756782897562757576852321402391681015249723862430503167634290721024776247393188943703605451071828541577890452008931414573470669791085941653286452972602880768000278121444139729224375672281655506532604713060219567780190942579824169994294072191071431007530810",
+ *          "rms":"61585968719825679820954752050482652604565722712299150828907711156378465728278927386099924570835148120166800215501590391175373817600340685471137391682056596396948247323267163942586709351681939970494142412129799511696342745194217108137415058975773547340948891027007892589530814649299044423034413987834238455096179370756258781816398285502424550467300576662889583917299579341488833013935611065055037617571364438681808991132527219626325739828132974032785684027848951273437930333539462921502518948344115455530438470689386576122673159249136945103964126302959119266362267068536535033404743724975610513609398570252057067331886",
+ *          "r":{
+ *              "sex":"2626632503410167078284139961009808200382092764903982154210447637569805753120300115008307139969073560497200188899592685451384854274242283626236697651644769767326232772277555517065937909501419381179080261665387996698800646130615385642112726371067011078728898840459575004146540068812617687191896173439535324651831638703293218277986742164291005504672543591242537338585952312913780174152675948770888046937709087672131923165052749838431320300824773077411638959605419049694924508977411338985002497282921570089360291100373020791726207566293384508958447187009585844585209052878418330808481074070751668704887853552386517055930",
+ *              "age":"63523998866502911303673581300753005318589707661733870330365925650678633010151708097940788515047278006579733539741582491459304064424323691384183804262052564249576046828258820357384995379735852197709005880259268104967579474816007639479449921766514022449488302232195920775417530914330757023422712527921767883202594148305685628110351091975069678465452002764246471559952603228501390812407272505516626682939968775033433977531801891624221994717435068960928494827798809830377239707582390852098331875021612002243884655607102948250212831978233087113810152800112957042083343256531406924266033890661377235591568062671765165489870",
+ *              "height":"36827765894725149703413262269832028490032908067763882745205554332012635847087233004524302605910728480570937402307712789502014630672100059296347324079688081495564484112324844430136540326393451614894312613497061339343299132067014791868694418918509496647956391816348038246819177729541875327561764868375504439504466305637343059834216402153387268518555623358163497146704705193066510197248636230544075701119933663421354632627652873131384367714145308220991601682858854828503784646138040614899662342567419634624453935944271936989617257057364382094088153891706740734234897521093567777490241869122832789895631974583798926624671",
+ *              "name":"18788372530355800304122414243192145025214535406665833599843110827678593803924867871859801890985253387426380024087621919226693609773752962328661361052659049213275281388746081914267713275478093781352622945304514254172068842629062418220350021713935789540517191405786855513356949833137385453147630963308137101514443239171812123677649126140681268375332458594132454494089548479553487688999824220296121412847356897601071257721654163708407527915287896754809415712208942681473722548922079792352789470647534972277610782248997591336611518792109685139554803952262932181423545140306161196124727474932461364805862202642858524609107"
+ *          },
+ *          "rctxt":"21056906618091809674328434216237046452978176268338751408357494760417916268345410074697140864434870457509245504931505200747385462128623958149802532264637271315245535230317207882554477612845326181656095735135034295869377522948921706594872112215344401196135236409557117610942403243622069726879719720245597225962847121582587787460228916727277232051837521756142362807695713742828849067021873709859012902116849601485209985334561870621322500146913090310536248902632690626159970312599495795517225254902488480051407151857092179309035985517180130778385441082409815097047032542548306682739499193653174348004839499978229431751109",
+ *          "z":"71880092276687608941911908974425364123157865039842259909133150431339154814634541364269068194300155348866964287780246139148914883270133559128984906638698935281508004126554770000460606903675678065228868678216890304285419036243351069244042246132289914582838445455253893803490784730193496852402692605579292635591572769896044739736701886904011628451851130193350093564432356884197344530964952157333799001881023464722691561680338381123701061260705675102730234526293501133803359541588132411602206704822105840321932467235097720843749769398614648948458076550887807392288231278953816644166369415087060316938918792123220256283592"
+ *      }
+ *  }
+ * }
+ */
+data class CredentialDefinition(
+    val ver: String,
+    val id: String,
+    @JsonProperty("schemaId") override val schemaId: String,
+    val type: String,
+    val tag: String,
+    val value: CredentialPubKeys,
+    @JsonIgnore override val credentialDefinitionId: String = id
+) : ContainsSchemaId, ContainsCredentialDefinitionId {
+    @JsonIgnore
+    fun getOwner() = id.split(":").first()
+
+    @JsonIgnore
+    fun getSchemaSeqNo() = schemaId
+
+    @JsonIgnore
+    fun getFilter() = """{schemaSeqNo:${getSchemaSeqNo()},owner:${getOwner()}}"""
+}
+
+data class CredentialPubKeys(
+    val primary: Any,
+    val revocation: Any?
+)
+
+/**
+ * Allows configure revocation when creating credential definition
+ */
+data class CredentialDefinitionConfig(val supportRevocation: Boolean = true)
+
+/**
+ * Credential definition id is the identifier of some credential definition. This class (de)serializes credential
+ * definition id. Indy uses raw string value of credential definition id, but it's parts are actually very useful.
+ * By possessing some credential definition id one could also know its:
+ *
+ * @param did: [String] - the DID of the credential definition issuer
+ * @param schemaSeqNo: [Int] - the schema's index on ledger
+ * @param tag: [String] - the tag of the credential definition (used for versioning)
+ */
+data class CredentialDefinitionId(val did: String, val schemaSeqNo: Int, val tag: String) {
+    override fun toString() = "$did:3:CL:$schemaSeqNo:$tag"
+
+    fun possibleRevocationRegistryDefinitionId(revTag: String)
+            = RevocationRegistryDefinitionId(did, this.toString(), revTag)
+
+    companion object : FromString<CredentialDefinitionId> {
+        override fun fromString(str: String): CredentialDefinitionId {
+            val strSplitted = str.split(":")
+
+            val didCred = strSplitted[0]
+            val tag = strSplitted[strSplitted.lastIndex]
+
+            val seqNo = strSplitted[3].toInt()
+
+            return CredentialDefinitionId(didCred, seqNo, tag)
+        }
+    }
+}
+
+/**
+ * Represents a class which somehow provides credential definition id
+ */
+interface ContainsCredentialDefinitionId {
+    val credentialDefinitionId: String
+    fun credentialDefinitionId() = CredentialDefinitionId.fromString(credentialDefinitionId)
+}
