+++ indy-utils/src/main/kotlin/com/luxoft/blockchainlab/hyperledger/indy/models/Proof.kt
@@ -1,44 +1,8 @@
-package com.luxoft.blockchainlab.hyperledger.indy
+package com.luxoft.blockchainlab.hyperledger.indy.models
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonProperty
-import org.hyperledger.indy.sdk.blob_storage.BlobStorageReader
-import org.hyperledger.indy.sdk.blob_storage.BlobStorageWriter
-
-/**
- * This file contains different data classes which are used to:
- *  1. Parse different indy responses from JSON
- *  2. Pass this data with some transport to other nodes
- */
-
-/**
- * Represents a JSON object with data we don't care about
- */
-typealias RawJsonMap = Map<String, Any?>
-
-/**
- * Timestamps in indy-world are represented as seconds from current unix epoch and are passed as ints.
- */
-object Timestamp {
-    fun now() = (System.currentTimeMillis() / 1000)
-}
-
-/**
- * Represents time interval used for non-revocation proof request creation
- */
-data class Interval(val from: Long?, val to: Long) {
-    companion object {
-        fun recent() = Interval(Timestamp.now() - 1, Timestamp.now())
-        fun allTime() = Interval(null, Timestamp.now())
-        fun now() = Interval(Timestamp.now(), Timestamp.now())
-    }
-}
-
-/**
- * Represents pairwise connection
- */
-data class ParsedPairwise(@JsonProperty("my_did") val myDid: String, val metadata: String)
 
 /**
  * Represents a particular attribute of a credential
@@ -54,155 +18,21 @@ data class CredentialFieldReference(
  */
 data class CredentialPredicate(val fieldReference: CredentialFieldReference, val value: Int, val type: String = ">=")
 
-/**
- * Represents credential offer structure from.
- *
- * Example:
- * {
- *  "schema_id":"V4SGRU86Z58d6TV7PBUe6f:2:gvt:1.0",
- *  "cred_def_id":"V4SGRU86Z58d6TV7PBUe6f:3:CL:11:TAG_1",
- *  "key_correctness_proof":{
- *      "c":"13761754840304977622723160610073451825885870027886575679644893270880210242048",
- *      "xz_cap":"340594989055020773127826593833336513921518724477036785882482817268296213948251287179407974588164919084334695651740569687300632847210154572884165159121599184025645166277098212790056340367675548818550765726296311292965992912074551019256115749936791678223911283791007656125545683865408985565241327271094579434170968145628655514446630544302951035920170794170246151362177816426545156184025449231122642094985676281843584455651701883820074327331609796847757062117491672731024235304103252859539661490806446048072318460507557965757317583074214165065297303144649946568723405884817293591612762145863174650181663890220148408024636352402091756427859989263831219418429810184080686385298141705690561349412826",
- *      "xr_cap":[
- *          [
- *              "sex",
- *              "114901712389925570747308755505959234181050635018455439015518673538665420714600742391112487936318767337811364541526998298909351153769381909405458975688912346829659002463105330151270376906904767852191120948221184076757548668178839902476764369158701651586213020361657048021989142635432083872470012807575132413533225131436609909265898568850849369239607732687307503724259039320052575044915702202127827497635261940052132289706849314498118894709987533054997486796495324835894021134546049747254662911845883745637179643031937372509664499727668214235022494858047366785583118269238559014142100466017577191051916007450284436475900646482814224888782960060518024693146673531063922330415915936725507733741766"
- *          ],
- *          [
- *              "height",
- *              "87925827250479498410024054128993506647408563496235591997576250953229036920959413598053360534336792807215061873972469226074015743490418917494435585457616775963853677170038454238754262365048434708885925169321027257392523529242823878080721156710414911339886319487461055502252343779360662670429306854831019261366623391128187478643887112210033870842054849023198347212729401623194059443628755320835685092944707226129817372493915681098131332323544477428002628840909180360070413137657346950550327191998701401766442452826340386823830760371963355720574367910987941501327853296595241416357764706345537910660613428363744193215240447807617629842317838672093122113424390775085560952953653914862062733460343"
- *          ],
- *          [
- *              "age",
- *              "293580247693151878814398198943884612452144653998669902194814683694462974264538931634280983530503958470285180952289780325087586570844097897587837512835758168930990380054580984461202240505574024403678079197953977922329186677771110248003920244900099391489674646294554422537185554996194781678780675220797160151412165442391881155725667489570391825773712503404857486672042567138673752180271219842671234810506725182196822255079859018162540163053327438979514404327868768200699692081177648933881745792483704320534292224956976674666455897176640767709821806484479564706964601579894630931579340752620471333071445922804710427962520331104667596003689846861532048740977647311084634223124206852506298920649196"
- *          ],
- *          [
- *              "name",
- *              "151838209223886649926918852654347303937974570263808192853823982025169602769061906064239685663594882187900023324484167008090681531752389241462722998187033577222372389532314754276217632204577207156930548234292944901257640439529710798080879389939943720570224719913036311420119996340541020890815695086382900217327755857286037846480798201129396307170715168897386494499064864127724739789806487609779050080617809107250172341879674283079820670267527232446397367776741773774314369698353651749919265249180651439166600686556843514441308902088097745692987067106281950141851140402550212265361047132303049470330028968862688869371904469006037728977048516991848156345476652473561273562256067871557885766170714"
- *          ]
- *      ]
- *  },
- *  "nonce":"107428647282355717425385"
- * }
- */
-data class CredentialOffer(
-    override val schemaId: String,
-    @JsonProperty("cred_def_id") override val credentialDefinitionId: String,
-    val keyCorrectnessProof: KeyCorrectnessProof,
-    val nonce: String
-) : ContainsSchemaId, ContainsCredentialDefinitionId
-
-data class KeyCorrectnessProof(val c: String, val xzCap: String, val xrCap: List<List<String>>)
 
 /**
- * Represents credential
- *
- * Example:
- * {
- *  "schema_id":"V4SGRU86Z58d6TV7PBUe6f:2:gvt:1.0",
- *  "cred_def_id":"V4SGRU86Z58d6TV7PBUe6f:3:CL:11:TAG_1",
- *  "rev_reg_id":null,
- *  "values":{
- *      "height":{
- *          "raw":"175",
- *          "encoded":"175"
- *      },
- *      "age":{
- *          "raw":"28",
- *          "encoded":"28"
- *      },
- *      "sex":{
- *          "raw":"male",
- *          "encoded":"5944657099558967239210949258394887428692050081607692519917050"
- *      },
- *      "name":{
- *          "raw":"Alex",
- *          "encoded":"1139481716457488690172217916278103335"
- *      }
- *   },
- *   "signature":{
- *      "p_credential":{
- *          "m_2":"49667387409072791895333303899794250195621079019939037173941098352432267351052",
- *          "a":"46472196767765770907394666564914908827636941437774002244808957817929335484901198469847816191888793787424454173615251713617437642140800957095444096953289022795933584947973882912549573017031326042172023742700124126097617123527165994497561684340831487031052756596925838552154440928523436704804747409373866015708822951087319456147879379870504295076305153309850751921996261053799107481550379520843249944485325939052177624821394027500110005624266210467198373597250503566237828450667266448183418227918400401704699885565245770811435243112403582975529471329394083897765235108123969958207741673911072934717461483113545266188644",
- *          "e":"259344723055062059907025491480697571938277889515152306249728583105665800713306759149981690559193987143012367913206299323899696942213235956742929778868433842729916830526494323897253",
- *          "v":"9338887988426201988765421562196770167154879800148014814333922180246755157926098177887019760590189644706588962889376430933272905256523280844850525307522972101041395604753489298645400791561173204918580803173446183370402135655019581889686304488882920336043712491329845108393523901867723835284305334314940293181692913962879688500165934035632199449508262321891885131213286008555035612673560763018079383317413701517586293134534116818845722224841680073553237282409082729185652870454427020765139523137039257689447228666527191123526642595385818366411298045778216509692039175062669997796006718498422877412133922090518369374952914510984290167002861717430237148314299826063537846309857330538893275349996073101096144565363217117483025547405594356909440065486323990781381766415423430083890878386623400226799579298704315850982500754033"
- *      },
- *      "r_credential":null
- *   },
- *   "signature_correctness_proof":{
- *      "se":"1284226342082548650319469664134606254016018076595188283130710346821644123498040330813630022863109638947088616640037389617195067147163300020258649281610103910536383626344523082298355692748130628416477490704057208576461724003588709297793772889042396317116869584149262342158061174817847081593704495774591524436555209826229614130157263249266720446677213266380692272756829476405786651218028670075213938724635127001673178988057648805807845178755090432507569406752220158843852268144659430034286675620223890177350461149637114343023508452878051585971310572205236878421406629892019776501637867344989734772211374366088132166423",
- *      "c":"27921508545451359693593235318800947503350694274033708495199465258660793720689"
- *   },
- *   "rev_reg":null,
- *   "witness":null
- *  }
+ * Represents credential reference
  */
-data class Credential(
-    override val schemaId: String,
-    @JsonProperty("cred_def_id") override val credentialDefinitionId: String,
-    @JsonProperty("rev_reg") val revocationRegistry: RawJsonMap?,
-    val witness: RawJsonMap?,
-    @JsonProperty("rev_reg_id") override val revocationRegistryId: String?,
-    val values: Map<String, CredentialValue>,
-    val signature: Map<String, RawJsonMap?>,
-    val signatureCorrectnessProof: RawJsonMap
-) : ContainsSchemaId, ContainsCredentialDefinitionId, ContainsRevocationRegistryId
-
-data class CredentialValue(val raw: String, val encoded: String)
-
-data class CredentialInfo(val credential: Credential, val credRevocId: String?, val revocRegDeltaJson: String?)
+data class ReferentCredential(val key: String, val credentialUUID: String)
 
 /**
- * Represents credential request
- *
- * Example:
- * {
- *  "prover_did":"V4SGRU86Z58d6TV7PBUe6f",
- *  "cred_def_id":"V4SGRU86Z58d6TV7PBUe6f:3:CL:11:TAG_1",
- *  "blinded_ms":{
- *      "u":"72450058149003836603428182664797985470713819794261103126480867221755083585000019083805880372220123137259404481119363367548533297964699411822665134867264734782748083131245644644829379611885825013696986984794316038880504276824857600130124722761268165098283024899944184232804885740644223950687056600506599750855615763905035549112546265746441459321427491710929699552131014192618377018001144050499899842399856735232631240922205358342612967266348360240558563773073817892531681534248845102456387131736087786422967038955678533370053125605715326022828981184036662477668484559656286399716565826498183240073734557435082596734065",
- *      "ur":null
- *   },
- *   "blinded_ms_correctness_proof":{
- *      "c":"31090608731660042115240737510118668008332292387792553721852538017836016343046",
- *      "v_dash_cap":"621718668163798225811486019897029182898804769156507569494749044241955862304516215318245510489009046007188401984864804588528641054499552984506583720030179227648767916980985982183552124476085689924812679856949430827708131983941887148948972968983789570929564392534906650079057667375456353133856194521441614464136439297585036819776218427302870626531718509167391460209553569722115239275775660807948979706555762042625362148149841737624635909956272979193386969558740798629192930114777368203501554600798530123840820679251140249079971589888320217707423350987128237788679490998464629247139577687505921078733467007002695480111134585335699005335287893428608708147970609552011478080930963461588079631276592063200753226925160329311",
- *      "ms_cap":"32160747770865816463105278561268942103905179895830400820850731916868271052319805671289486219845889897290386865499398383995979398737576225013066455933796815223495239463836014358141"
- *   },
- *   "nonce":"1186917666793748761028490"
- * }
- *
+ * Represents data which is needed for verifier to verify proof
+ * Data in this data class is stored as JSON
  */
-data class CredentialRequestInfo(
-    val request: CredentialRequest,
-    val metadata: CredentialRequestMetadata
-)
-
-data class CredentialRequest(
-    val proverDid: String,
-    @JsonProperty("cred_def_id") override val credentialDefinitionId: String,
-    val blindedMs: RawJsonMap,
-    val blindedMsCorrectnessProof: RawJsonMap,
-    val nonce: String
-) : ContainsCredentialDefinitionId
-
-/**
- * Represents credential request metadata
- *
- * Example:
- * {
- *  "master_secret_blinding_data":{
- *      "v_prime":"19996992452923335739444111602323353146400610615332675413132845150786968475048282265691890632536001360833725249574222779467320862724682652038419474915879533807321016579521760714569205647449244233907560466688603170418829351681403404935647974607945749195678658844745904385566324446192755479942309264166454465738364166358492528736414247799673112394278935683056028326248487491960209103589894282360872956217377481736453946808616539396839604132182798890680363804230266296163940604353270837853010999758985713684313386130790502196378124744035357918753867699384331285330409134810007357798671352898872667892627983510765893834524972207798318482093594850",
- *      "vr_prime":null
- *  },
- *  "nonce":"1186917666793748761028490",
- *  "master_secret_name":"masterSecretId"
- * }
- */
-data class CredentialRequestMetadata(
-    val masterSecretBlindingData: RawJsonMap,
-    val masterSecretName: String,
-    val nonce: String
+data class DataUsedInProofJson(
+    val schemas: String,
+    val credentialDefinitions: String,
+    val revocationRegistryDefinitions: String,
+    val revocationRegistries: String
 )
 
 /**
@@ -612,339 +442,3 @@ data class RequestedProof(
 )
 
 data class ProofDetails(val primaryProof: Any, @JsonProperty("non_revoc_proof") val nonRevokedProof: Any?)
-
-/**
- * Represents indy schema
- *
- * @param id                identifier of schema
- * @param attributeNames    an array of attribute name strings
- * @param seqNo             ??? Int
- * @param name              Schema's name string
- * @param version           Schema's version string
- * @param ver              Version of the Schema json
- */
-data class Schema(
-    val ver: String,
-    val id: String,
-    val name: String,
-    val version: String,
-    @JsonProperty("attrNames") val attributeNames: List<String>,
-    @JsonProperty("seqNo") val seqNo: Int?,
-    @JsonIgnore override val schemaId: String = id
-) : ContainsSchemaId {
-    @JsonIgnore
-    fun getOwner() = id.split(":").first()
-
-    @JsonIgnore
-    fun isValid() = seqNo != null
-
-    @JsonIgnore
-    fun getFilter() = """{name:$name,version:$version,owner:${getOwner()}}"""
-}
-
-/**
- * Represents credential definition
- *
- * Example:
- * {
- *  "ver":"1.0",
- *  "id":"V4SGRU86Z58d6TV7PBUe6f:3:CL:11:TAG_1",
- *  "schemaId":"11",
- *  "type":"CL",
- *  "tag":"TAG_1",
- *  "value":{
- *      "primary":{
- *          "n":"97825789590062692150319390633467047046201095546380339364160961030429137551345980953797290715661332171822654941232742526180523305975436127060967945697600291470349688931509035153112372164175391196321435197412518123214148508985637308223324599142177695725198704394816353613227336593431540982906270549568391784047827941399905911323367262453292590513881444395762311018337780680352564520585183115243471839552585387700359266104870621499920269868340747617483585174509821630813967372030256014647503731893474943716949401787092423796309347202300689031224683810152043236004828131092812383226231107878482692717606522179435737963809",
- *          "s":"39733411017179560244024968899642664398180361222127496719414627387111951836383304903709718151952941019469808679723550262854530221428192476695197984887679496172637317619920757202465712652277595749798862290700699750698253379509815941232001323375550623145841707962772398431536953095683812602550974096101709249765431308450233784849698817513922831987542597248666167913756782897562757576852321402391681015249723862430503167634290721024776247393188943703605451071828541577890452008931414573470669791085941653286452972602880768000278121444139729224375672281655506532604713060219567780190942579824169994294072191071431007530810",
- *          "rms":"61585968719825679820954752050482652604565722712299150828907711156378465728278927386099924570835148120166800215501590391175373817600340685471137391682056596396948247323267163942586709351681939970494142412129799511696342745194217108137415058975773547340948891027007892589530814649299044423034413987834238455096179370756258781816398285502424550467300576662889583917299579341488833013935611065055037617571364438681808991132527219626325739828132974032785684027848951273437930333539462921502518948344115455530438470689386576122673159249136945103964126302959119266362267068536535033404743724975610513609398570252057067331886",
- *          "r":{
- *              "sex":"2626632503410167078284139961009808200382092764903982154210447637569805753120300115008307139969073560497200188899592685451384854274242283626236697651644769767326232772277555517065937909501419381179080261665387996698800646130615385642112726371067011078728898840459575004146540068812617687191896173439535324651831638703293218277986742164291005504672543591242537338585952312913780174152675948770888046937709087672131923165052749838431320300824773077411638959605419049694924508977411338985002497282921570089360291100373020791726207566293384508958447187009585844585209052878418330808481074070751668704887853552386517055930",
- *              "age":"63523998866502911303673581300753005318589707661733870330365925650678633010151708097940788515047278006579733539741582491459304064424323691384183804262052564249576046828258820357384995379735852197709005880259268104967579474816007639479449921766514022449488302232195920775417530914330757023422712527921767883202594148305685628110351091975069678465452002764246471559952603228501390812407272505516626682939968775033433977531801891624221994717435068960928494827798809830377239707582390852098331875021612002243884655607102948250212831978233087113810152800112957042083343256531406924266033890661377235591568062671765165489870",
- *              "height":"36827765894725149703413262269832028490032908067763882745205554332012635847087233004524302605910728480570937402307712789502014630672100059296347324079688081495564484112324844430136540326393451614894312613497061339343299132067014791868694418918509496647956391816348038246819177729541875327561764868375504439504466305637343059834216402153387268518555623358163497146704705193066510197248636230544075701119933663421354632627652873131384367714145308220991601682858854828503784646138040614899662342567419634624453935944271936989617257057364382094088153891706740734234897521093567777490241869122832789895631974583798926624671",
- *              "name":"18788372530355800304122414243192145025214535406665833599843110827678593803924867871859801890985253387426380024087621919226693609773752962328661361052659049213275281388746081914267713275478093781352622945304514254172068842629062418220350021713935789540517191405786855513356949833137385453147630963308137101514443239171812123677649126140681268375332458594132454494089548479553487688999824220296121412847356897601071257721654163708407527915287896754809415712208942681473722548922079792352789470647534972277610782248997591336611518792109685139554803952262932181423545140306161196124727474932461364805862202642858524609107"
- *          },
- *          "rctxt":"21056906618091809674328434216237046452978176268338751408357494760417916268345410074697140864434870457509245504931505200747385462128623958149802532264637271315245535230317207882554477612845326181656095735135034295869377522948921706594872112215344401196135236409557117610942403243622069726879719720245597225962847121582587787460228916727277232051837521756142362807695713742828849067021873709859012902116849601485209985334561870621322500146913090310536248902632690626159970312599495795517225254902488480051407151857092179309035985517180130778385441082409815097047032542548306682739499193653174348004839499978229431751109",
- *          "z":"71880092276687608941911908974425364123157865039842259909133150431339154814634541364269068194300155348866964287780246139148914883270133559128984906638698935281508004126554770000460606903675678065228868678216890304285419036243351069244042246132289914582838445455253893803490784730193496852402692605579292635591572769896044739736701886904011628451851130193350093564432356884197344530964952157333799001881023464722691561680338381123701061260705675102730234526293501133803359541588132411602206704822105840321932467235097720843749769398614648948458076550887807392288231278953816644166369415087060316938918792123220256283592"
- *      }
- *  }
- * }
- */
-data class CredentialDefinition(
-    val ver: String,
-    val id: String,
-    @JsonProperty("schemaId") override val schemaId: String,
-    val type: String,
-    val tag: String,
-    val value: CredentialPubKeys,
-    @JsonIgnore override val credentialDefinitionId: String = id
-) : ContainsSchemaId, ContainsCredentialDefinitionId {
-    @JsonIgnore
-    fun getOwner() = id.split(":").first()
-
-    @JsonIgnore
-    fun getSchemaSeqNo() = schemaId
-
-    @JsonIgnore
-    fun getFilter() = """{schemaSeqNo:${getSchemaSeqNo()},owner:${getOwner()}}"""
-}
-
-data class CredentialPubKeys(
-    val primary: Any,
-    val revocation: Any?
-)
-
-data class RevocationRegistryConfig(
-    val issuanceType: String,
-    @JsonProperty("max_cred_num") val maximumCredentialNumber: Int
-)
-
-/**
- * Represents revocation registry definition
- *
- * Example:
- * {
- *  "ver":"1.0",
- *  "id":"V4SGRU86Z58d6TV7PBUe6f:4:V4SGRU86Z58d6TV7PBUe6f:3:CL:11:TAG_1:CL_ACCUM:REV_TAG_1",
- *  "revocDefType":"CL_ACCUM",
- *  "tag":"REV_TAG_1",
- *  "credDefId":"V4SGRU86Z58d6TV7PBUe6f:3:CL:11:TAG_1",
- *  "value":{
- *      "issuanceType":"ISSUANCE_ON_DEMAND",
- *      "maxCredNum":10000,
- *      "publicKeys":{
- *          "accumKey":{
- *              "z":"DD5CB7C7B73632 4AFF49D65DC22B 75DCE6B4720E9A 6850C5B997857B 24A81D0A FB91DEDCC2933B 58F8253DDC2932 A70370A1A6B790 B7C1D0EA96C211 1F9AC413 42BCA194D89D6E 2CA77CB3C7D22A 5E2004C628FE02 E3AF249480D877 1078CDD3 1899F0C8F69EF0 6E6597A07EFCBE 3050DA53AC48F 138D31F5D0F836 20DD73AA E8CBC1334EAC3E 6221F7D1C21FBC AB5605E23860D7 BBF7B256371799 2E756F8 C7ECC90D700DAB D1A7EEED09CB33 DA218E8EC0C2E7 93EC5FF2FE457C 1861FA59 63FB8BC55D915 B726AE490AC56C 49ED7DEFC0988D 60815A62EC29CD 1D1E8504 76C9A801569840 5E417CE5540DCD 77FCEDF0A5DD9 47D9C0D070C4B1 23315D95 87773524083058 E75B5A54FF24F 33148931C2BB4E 426BBE4DC6AA66 3C910CC 66C14B91B5D70A FD94681339A7B5 D27A926A28D6AE 385A898772ED98 797FFA8 5A591894E431CF 582624540B9B0C 28E11C07575D81 8D96EE6F5EFB27 16D6BD9B DF160431970E42 3AAE8325F8F8B8 93D8C65022890A 485AFA07D3F281 423B31C"
- *          }
- *      },
- *      "tailsHash":"FU6TF1Tw8D2Pk8MT8y5DVZSBUJqq3WdGYvTm3oGU2hYS",
- *      "tailsLocation":"/home/joinu/.indy_client/tails/FU6TF1Tw8D2Pk8MT8y5DVZSBUJqq3WdGYvTm3oGU2hYS"
- *  }
- * }
- */
-
-data class RevocationRegistryInfo(
-    val definition: RevocationRegistryDefinition,
-    val entry: RevocationRegistryEntry
-)
-
-data class RevocationRegistryDefinition(
-    val ver: String,
-    val id: String,
-    @JsonProperty("revocDefType") val revocationRegistryDefinitionType: String,
-    val tag: String,
-    @JsonProperty("credDefId") override val credentialDefinitionId: String,
-    val value: RawJsonMap,
-    @JsonIgnore override val revocationRegistryId: String? = id
-) : ContainsCredentialDefinitionId, ContainsRevocationRegistryId
-
-/**
- * Represents revocation registry entry
- *
- * Example
- * {
- *  "ver":"1.0",
- *  "value":{
- *      "accum":"true 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0"
- *  }
- * }
- */
-data class RevocationRegistryEntry(
-    val ver: String,
-    val value: RawJsonMap
-)
-
-data class RevocationState(
-    val witness: RawJsonMap,
-    @JsonProperty("rev_reg") val revocationRegistry: RawJsonMap,
-    val timestamp: Long,
-    @JsonIgnore override var revocationRegistryId: String? = null
-) : ContainsRevocationRegistryId
-
-/**
- * Abstracts blob storage reader and writer which are used for tails file management
- */
-data class BlobStorageHandler(val reader: BlobStorageReader, val writer: BlobStorageWriter)
-
-/**
- * Represents credential reference
- */
-data class ReferentCredential(val key: String, val credentialUUID: String)
-
-/**
- * Represents data which is needed for verifier to verify proof
- * Data in this data class is stored as JSON
- */
-data class DataUsedInProofJson(
-    val schemas: String,
-    val credentialDefinitions: String,
-    val revocationRegistryDefinitions: String,
-    val revocationRegistries: String
-)
-
-data class StorageConfig(val path: String)
-
-/**
- * {
- *     "id": string, Identifier of the wallet. Configured storage uses this identifier to lookup exact wallet data placement.
- *
- *     "storage_type": optional<string>, Type of the wallet storage. Defaults to 'default'.
- *     'Default' storage type allows to store wallet data in the local file.
- *     Custom storage types can be registered with indy_register_wallet_storage call.
- *
- *     "storage_config": optional<object>, Storage configuration json. Storage type defines set of supported keys.
- *     Can be optional if storage supports default configuration.
- *
- *     For 'default' storage type configuration is:
- *     {
- *         "path": optional<string>, Path to the directory with wallet files.
- *         Defaults to $HOME/.indy_client/wallets.
- *         Wallet will be stored in the file {path}/{id}/sqlite.db
- *     }
- * }
- */
-data class WalletConfig(
-    val id: String,
-    val storageType: String = "default",
-    val storageConfig: StorageConfig? = null
-)
-
-/**
- * Represents some details of a particular identity
- *
- * @param did: [String]             did of this identity
- * @param verkey: [String]          verification key of this identity
- * @param alias: [String]           <optional> additional alias of this identity
- * @param role: [String]            <optional> role of this identity (e.g. TRUSTEE)
- */
-data class IdentityDetails(
-    val did: String,
-    val verkey: String,
-    @JsonIgnore val alias: String?,
-    @JsonIgnore val role: String?
-) {
-    @JsonIgnore
-    fun getIdentityRecord() = """{"did":"$did","verkey":"$verkey"}"""
-}
-
-/**
- * Schema id is the local identifier of some schema. Don't confuse with schema sequence number which represents schema's
- * index on ledger. This class (de)serializes schema id.
- * Indy uses raw string value of schema id, but it's parts are actually very useful. By possessing some schema id one
- * could also know its:
- *
- * @param did: [String] - the DID of the schema issuer
- * @param name: [String] - the name of the schema
- * @param version: [String] - the version of the schema
- */
-class SchemaId(val did: String, val name: String, val version: String) {
-    override fun toString() = "$did:2:$name:$version"
-
-    companion object : FromString<SchemaId> {
-        override fun fromString(str: String): SchemaId {
-            val (did, _, name, version) = str.split(":")
-
-            return SchemaId(did, name, version)
-        }
-    }
-}
-
-/**
- * Credential definition id is the identifier of some credential definition. This class (de)serializes credential
- * definition id. Indy uses raw string value of credential definition id, but it's parts are actually very useful.
- * By possessing some credential definition id one could also know its:
- *
- * @param did: [String] - the DID of the credential definition issuer
- * @param schemaSeqNo: [Int] - the schema's index on ledger
- * @param tag: [String] - the tag of the credential definition (used for versioning)
- */
-data class CredentialDefinitionId(val did: String, val schemaSeqNo: Int, val tag: String) {
-    override fun toString() = "$did:3:CL:$schemaSeqNo:$tag"
-
-    fun possibleRevocationRegistryDefinitionId(revTag: String)
-            = RevocationRegistryDefinitionId(did, this.toString(), revTag)
-
-    companion object : FromString<CredentialDefinitionId> {
-        override fun fromString(str: String): CredentialDefinitionId {
-            val strSplitted = str.split(":")
-
-            val didCred = strSplitted[0]
-            val tag = strSplitted[strSplitted.lastIndex]
-
-            val seqNo = strSplitted[3].toInt()
-
-            return CredentialDefinitionId(didCred, seqNo, tag)
-        }
-    }
-}
-
-/**
- * Revocation registry definition id is the local identifier of some revocation registry. This class (de)serializes
- * revocation registry definition id.
- * Indy uses raw string value of revocation registry definition id, but it's parts are actually very useful.
- * By possessing some revocation registry definition id id one could also know its:
- *
- * @param did: [String] - the DID of the revocation registry issuer
- * @param credentialDefinitionId: [CredentialDefinitionId] - the id of credential definition which
- *  this revocation registry accompanies
- * @param tag: [String] - the tag of the revocation registry definition (used for versioning)
- */
-data class RevocationRegistryDefinitionId(
-    val did: String,
-    override val credentialDefinitionId: String,
-    val tag: String
-): ContainsCredentialDefinitionId {
-    override fun toString() = "$did:4:$credentialDefinitionId:CL_ACCUM:$tag"
-
-    companion object : FromString<RevocationRegistryDefinitionId> {
-        override fun fromString(str: String): RevocationRegistryDefinitionId {
-            val strSplitted = str.split(":")
-            val didRev = strSplitted[0]
-            val tagRev = strSplitted[strSplitted.lastIndex]
-            val didCred = strSplitted[2]
-            val tagCred = strSplitted[strSplitted.lastIndex - 2]
-
-            val seqNo = strSplitted[5].toInt()
-
-            return RevocationRegistryDefinitionId(didRev, CredentialDefinitionId(didCred, seqNo, tagCred).toString(), tagRev)
-        }
-    }
-}
-
-/**
- * Interface for class that can be constructed from some string data
- */
-interface FromString<T : Any> {
-    fun fromString(str: String): T
-}
-
-/**
- * Represents a class which somehow provides schema id
- */
-interface ContainsSchemaId {
-    val schemaId: String
-    fun schemaId() = SchemaId.fromString(schemaId)
-}
-
-/**
- * Represents a class which somehow provides credential definition id
- */
-interface ContainsCredentialDefinitionId {
-    val credentialDefinitionId: String
-    fun credentialDefinitionId() = CredentialDefinitionId.fromString(credentialDefinitionId)
-}
-
-/**
- * Represents a class which somehow provides revocation registry definition id
- */
-interface ContainsRevocationRegistryId {
-    val revocationRegistryId: String?
-    fun revocationRegistryId() =
-        if (revocationRegistryId == null) null
-        else RevocationRegistryDefinitionId.fromString(revocationRegistryId!!)
-}
-
-data class WalletPassword(val key: String)
-
-data class TailsConfig(val baseDir: String, val uriPattern: String = "")
-
-data class CredentialDefinitionConfig(val supportRevocation: Boolean = true)
