+++ indy-utils/src/main/kotlin/com/luxoft/blockchainlab/hyperledger/indy/roles/IndyTrustee.kt
@@ -1,6 +1,6 @@
 package com.luxoft.blockchainlab.hyperledger.indy.roles
 
-import com.luxoft.blockchainlab.hyperledger.indy.IdentityDetails
+import com.luxoft.blockchainlab.hyperledger.indy.models.IdentityDetails
 
 
 /**
