+++ indy-utils/src/main/kotlin/com/luxoft/blockchainlab/hyperledger/indy/roles/IndyVerifier.kt
@@ -1,6 +1,6 @@
 package com.luxoft.blockchainlab.hyperledger.indy.roles
 
-import com.luxoft.blockchainlab.hyperledger.indy.*
+import com.luxoft.blockchainlab.hyperledger.indy.models.*
 import org.hyperledger.indy.sdk.pool.Pool
